Refactored to split the Controller into three Controllers each with a more narrow purpose
